Finalizado, con cabecera y permientiendo doubles

+++ viajanteGit/src/viajantegit/Nodo.java
@@ -8,14 +8,14 @@ package viajantegit;
 
 public class Nodo 
 {
-	private int x;
-	private int y;
+	private double x;
+	private double y;
         private int pos;
         
         private int valor;
 	
 	
-	public Nodo(int pos, int x, int y) 
+	public Nodo(int pos, double x, double y) 
 	{
 		super();
                 this.pos = pos;
@@ -31,13 +31,13 @@ public class Nodo
             this.y = nodo.y;
         }
         
-	public int getX() {
+	public double getX() {
 		return x;
 	}
 	public void setX(int x) {
 		this.x = x;
 	}
-	public int getY() {
+	public double getY() {
 		return y;
 	}
 	public void setY(int y) {
